Attempted fix of memory leak
There was a memory leak in the listenRedisStream.ts file, I've attempted to fix it by moving some variable declarations to be outside the while loop. This should reduce the number of listeners on the abort signal. Needs testing to verify.
Also fixed a small problem with the content security policy and the event card factory handling of null values.

Signed-off-by: Brian Evans <ebrian101@gmail.com>