Count resource kinds by stream. Changed keys.
Keep a tally of how many of each resource kind come on each stream. This will be useful to pick schemas (which are saved by resource kind).
Also in this commit: changed the timepoint and heartbeat keys to be a redis hash instead to reduce the number of keys and store related information together.

Signed-off-by: Brian Evans <ebrian101@gmail.com>